requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.2.1
discord-py==2.6.4
eventkit==1.0.3
executing==2.2.1
frozenlist==1.5.0
ib-insync==0.9.86
idna==3.11
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.24.4
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.52
propcache==0.2.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
python-dotenv==1.0.1
stack-data==0.6.3
tomli==2.3.0
traitlets==5.14.3
typing-extensions==4.13.2
wcwidth==0.2.14
yarl==1.15.2
